service/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
celery[redis]==5.2.7
celery_singleton==0.3.1
